+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add self-reference `craneLib` to crane lib instance.
 * Add `removeReferencesToRustToolchainHook`, which, by default, removes
   references to the Rust toolchain from all installed binaries
+* `mkCargoDerivation` now configures the cross-compilation toolchain by default
+  using the newly added `mkCrossToolchainEnv` function. To select the utilized
+  cross compiler, the `stdenv` argument now also accepts a selector function,
+  taking an arbitrary `pkgs` instantiation as an argument. This behavior can be
+  turned off using the `noCrossToolchainEnv` argument.
   
 ### Changed
 * `buildPackage` now includes `removeReferencesToRustToolchainHook` as a native